requirements.txt
Django
djangorestframework
pillow
psycopg2-binary
python-decouple
drf-yasg
django-filter